CNV 19687: SidebarEditor wizard other tabs
