--- src/views/catalog/wizard/components/WizardDescriptionItem/WizardDescriptionItem.scss
@@ -1,4 +1,0 @@
-.wizard-description-item__title {
-  justify-content: space-between;
-  flex: 1;
-}
