+++ src/views/catalog/wizard/tabs/disks/wizard-disk-tab.scss
@@ -1,0 +1,12 @@
+.wizard-disk-tab {
+    height: 100%;
+
+    .co-m-pane__body {
+        height: 100%;
+    }
+
+
+    .pf-c-sidebar__panel {
+        padding-top: var(--pf-global--spacer--md);
+    }
+}
