+++ src/views/catalog/wizard/tabs/metadata/wizard-metadata-tab.scss
@@ -1,0 +1,13 @@
+.wizard-metadata-tab {
+
+    .wizard-description-item__title {
+        justify-content: space-between;
+        flex: 1;
+      }
+
+    &__description-list {
+        max-width: 700px;
+        margin-top: var(--pf-global--spacer--md);
+    }
+      
+}
