+++ src/views/catalog/wizard/tabs/network/WizardNetworkTab.tsx
@@ -2,13 +2,17 @@ import * as React from 'react';
 
 import { WizardTab } from '@catalog/wizard/tabs';
 import { useModal } from '@kubevirt-utils/components/ModalProvider/ModalProvider';
+import SidebarEditor from '@kubevirt-utils/components/SidebarEditor/SidebarEditor';
+import SidebarEditorSwitch from '@kubevirt-utils/components/SidebarEditor/SidebarEditorSwitch';
 import { useKubevirtTranslation } from '@kubevirt-utils/hooks/useKubevirtTranslation';
 import { ListPageBody, ListPageCreateButton } from '@openshift-console/dynamic-plugin-sdk';
+import { Flex, FlexItem } from '@patternfly/react-core';
 
 import NetworkInterfaceList from './components/list/NetworkInterfaceList';
 import NetworkInterfaceModal from './components/modal/NetworkInterfaceModal';
 
 import 'src/utils/styles/ListPageCreateButton.scss';
+import './wizard-network-tab.scss';
 
 const WizardNetworkTab: WizardTab = ({ vm, updateVM }) => {
   const { t } = useKubevirtTranslation();
@@ -16,27 +20,36 @@ const WizardNetworkTab: WizardTab = ({ vm, updateVM }) => {
 
   const actionText = t('Add network interface');
   return (
-    <>
+    <div className="wizard-network-tab">
       <ListPageBody>
-        <ListPageCreateButton
-          className="list-page-create-button-margin"
-          onClick={() =>
-            createModal(({ isOpen, onClose }) => (
-              <NetworkInterfaceModal
-                isOpen={isOpen}
-                onClose={onClose}
-                headerText={actionText}
-                vm={vm}
-                updateVM={updateVM}
-              />
-            ))
-          }
-        >
-          {actionText}
-        </ListPageCreateButton>
-        <NetworkInterfaceList vm={vm} />
+        <SidebarEditor resource={vm} onResourceUpdate={(newVM) => updateVM(newVM)}>
+          <Flex>
+            <FlexItem>
+              <ListPageCreateButton
+                className="list-page-create-button-margin"
+                onClick={() =>
+                  createModal(({ isOpen, onClose }) => (
+                    <NetworkInterfaceModal
+                      isOpen={isOpen}
+                      onClose={onClose}
+                      headerText={actionText}
+                      vm={vm}
+                      updateVM={updateVM}
+                    />
+                  ))
+                }
+              >
+                {actionText}
+              </ListPageCreateButton>
+            </FlexItem>
+            <FlexItem>
+              <SidebarEditorSwitch />
+            </FlexItem>
+          </Flex>
+          <NetworkInterfaceList vm={vm} />
+        </SidebarEditor>
       </ListPageBody>
-    </>
+    </div>
   );
 };
