+++ src/views/catalog/wizard/tabs/scheduling/wizard-scheduling-tab.scss
@@ -1,0 +1,6 @@
+
+.wizard-scheduling-tab {
+    &__grid {
+      margin-top: var(--pf-global--spacer--md);
+    }
+  }
