+++ src/views/catalog/wizard/tabs/scripts/WizardScriptsTab.scss
@@ -1,8 +1,16 @@
-.wizard-scripts-tab-expanable-section {
-  .pf-c-expandable-section__toggle {
-    align-items: center;
+.wizard-scripts-tab {
+
+  &__description-list {
+    max-width: 700px;
+    margin-top: var(--pf-global--spacer--md);
   }
-  .pf-c-expandable-section__toggle-text {
-    font-size: 18px;
+  
+  &-expanable-section {
+    .pf-c-expandable-section__toggle {
+      align-items: center;
+    }
+    .pf-c-expandable-section__toggle-text {
+      font-size: var(--pf-global--FontSize--lg);
+    }
   }
 }
