+++ src/views/templates/details/tabs/details/TemplateDetailsPage.scss
@@ -4,7 +4,7 @@
 }
 
 .margin-top {
-  margin-top: 1rem;
+  margin-top: var(--pf-global--spacer--md);
 }
 
 .title-edit-item-space {
